Auto Pilot - no control
Starting to implement an auto piloting system.

Trying this with classes and a doubly linked list in a separate .h / .cpp to keep it clean.
Not quite sure about what functions are needed to manage this, but I don't worry about that yet.

Also changed tthe data structure for the nodes. Instead of an action and a value for that action it's only one value/command now.